Move matrix multiplication into a separate crate
Extract the matrix multiplication implementation from the main crate
into a new rten-gemm crate.
