+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
   ".",
   "rten-base",
   "rten-cli",
+  "rten-gemm",
   "rten-generate",
   "rten-imageio",
   "rten-imageproc",
@@ -49,6 +50,7 @@ include = ["/src", "/CHANGELOG.md", "/README.md"]
 flatbuffers = "24.3.25"
 smallvec = { version = "1.10.0", features = ["union", "const_generics", "const_new"] }
 rten-base = { path = "./rten-base" }
+rten-gemm = { path = "./rten-gemm" }
 rten-tensor = { path = "./rten-tensor", version = "0.20.0" }
 rten-vecmath = { path = "./rten-vecmath", version = "0.19.0" }
 rten-simd = { path = "./rten-simd", version = "0.19.0" }
