+++ rten-gemm/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "rten-gemm"
+version = "0.20.0"
+edition = "2021"
+authors = ["Robert Knight"]
+description = "Machine-learning oriented matrix multiplication"
+license = "MIT or Apache-2.0"
+homepage = "https://github.com/robertknight/rten"
+repository = "https://github.com/robertknight/rten"
+
+[dependencies]
+rayon = { workspace = true }
+rten-base = { path = "../rten-base" }
+rten-simd = { path = "../rten-simd" }
+rten-tensor = { path = "../rten-tensor" }
+
+[dev-dependencies]
+rten-bench = { path = "../rten-bench" }
+rten-testing = { path = "../rten-testing" }
+
+[features]
+avx512 = ["rten-simd/avx512"]
+
+[lib]
+crate-type = ["lib"]
+
+[lints.clippy]
+# Clippy frequently suggests to replace for loops with const bounds (often used
+# in performance-critical loops) with iterators, which is more verbose and
+# potentially less efficient.
+needless_range_loop = "allow"
+too_many_arguments = "allow"
+uninlined_format_args = "allow"
