+++ rten-gemm/README.md
@@ -1,0 +1,4 @@
+# rten-gemm
+
+This crate contains optimized matrix-multiplication implementations, for both
+floating point and int8 inputs.
