+++ src/lib.rs
@@ -116,7 +116,6 @@ use rten_tensor::{NdTensor, Tensor};
 mod constant_storage;
 mod downcast;
 mod env;
-mod gemm;
 mod graph;
 mod header;
 mod model;
